requirements.txt
requests
atlassian-python-api
essential_generators
python-dotenv
random-word
pyyaml